+++ exceptions.py
@@ -1,0 +1,11 @@
+"Определение пользовательских исключений."
+
+
+class MissingSourceError(Exception):
+    """Класс представляет ошибку, возникающую при отсутствии источника изображения."""
+    pass
+
+
+class TwoSourcesError(Exception):
+    """Класс представляет ошибку, возникающую при указании двух источников изображений."""
+    pass
